Add Basic Table

+++ be-artwork-sharing-platform/Core/Entities/Category.cs
@@ -3,7 +3,7 @@
 namespace be_artwork_sharing_platform.Core.Entities
 {
     [Table("categories")]
-    public class Category : BaseEntity<int>
+    public class Category : BaseEntity<long>
     {
         public string Name { get; set; }
     }
